LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-units-animation-expected.txt


FAIL Animation using cqw unit assert_equals: expected "60px" but got "auto"
FAIL Animation using cqw unit responds to changing container size assert_equals: expected "60px" but got "auto"
FAIL Animation using cqh unit assert_equals: expected "60px" but got "auto"
FAIL Animation using cqh unit responds to changing container size assert_equals: expected "60px" but got "auto"
FAIL Animation using cqi unit assert_equals: expected "60px" but got "auto"
FAIL Animation using cqi unit responds to changing container size assert_equals: expected "60px" but got "auto"
FAIL Animation using cqb unit assert_equals: expected "60px" but got "auto"
FAIL Animation using cqb unit responds to changing container size assert_equals: expected "60px" but got "auto"
FAIL Animation using cqmin unit assert_equals: expected "60px" but got "auto"
FAIL Animation using cqmin unit responds to changing container size assert_equals: expected "60px" but got "auto"
FAIL Animation using cqmax unit assert_equals: expected "60px" but got "auto"
FAIL Animation using cqmax unit responds to changing container size assert_equals: expected "60px" but got "auto"
